Incidents-data-scrapper/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0

anthropic

torch>=2.1.0
datasets>=2.14.0
transformers>=4.36.0
trl>=0.7.0
unsloth
